refactor: create javascript helper
